requirements.txt
python-telegram-bot
pytest
requests